feat: Implemented most of the service endpoints already

+++ mynd-backend/src/main/java/de/thi/mynd/auth/dto/PersonalInvitationStatusDto.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 @Builder
 public final class PersonalInvitationStatusDto {
     public int invitationsLeft;
-    public int invitationsAlreadySent;
+    public long invitationsAlreadySent;
 }
